GTPSE-36: Refectoring module
