--- modules/amf-registration/amf-registration-rest/amf-registration-rest-api/src/main/java/com/liferay/amf/resource/v1_0/AccountResource.java
@@ -1,140 +1,0 @@
-package com.liferay.amf.resource.v1_0;
-
-import com.liferay.amf.dto.v1_0.Account;
-import com.liferay.portal.kernel.search.filter.Filter;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.ResourceActionLocalService;
-import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.odata.filter.ExpressionConvert;
-import com.liferay.portal.odata.filter.FilterParserProvider;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-import com.liferay.portal.vulcan.pagination.Page;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-
-import javax.annotation.Generated;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
-
-import org.osgi.annotation.versioning.ProviderType;
-
-/**
- * To access this resource, run:
- *
- *     curl -u your@email.com:yourpassword -D - http://localhost:8080/o/amf-registration-rest/v1.0
- *
- * @author me
- * @generated
- */
-@Generated("")
-@ProviderType
-public interface AccountResource {
-
-	public static Builder builder() {
-		return FactoryHolder.factory.create();
-	}
-
-	public Page<Account> getAllAccount(Integer pageNumber, Integer pageSize)
-		throws Exception;
-
-	public Account createAccount(Account account) throws Exception;
-
-	public void deleteAccount(String accountId) throws Exception;
-
-	public Response deleteAccountBatch(String callbackURL, Object object)
-		throws Exception;
-
-	public Account getAccount(String accountId) throws Exception;
-
-	public Account updateAccount(String accountId, Account account)
-		throws Exception;
-
-	public default void setContextAcceptLanguage(
-		AcceptLanguage contextAcceptLanguage) {
-	}
-
-	public void setContextCompany(
-		com.liferay.portal.kernel.model.Company contextCompany);
-
-	public default void setContextHttpServletRequest(
-		HttpServletRequest contextHttpServletRequest) {
-	}
-
-	public default void setContextHttpServletResponse(
-		HttpServletResponse contextHttpServletResponse) {
-	}
-
-	public default void setContextUriInfo(UriInfo contextUriInfo) {
-	}
-
-	public void setContextUser(
-		com.liferay.portal.kernel.model.User contextUser);
-
-	public void setExpressionConvert(
-		ExpressionConvert<Filter> expressionConvert);
-
-	public void setFilterParserProvider(
-		FilterParserProvider filterParserProvider);
-
-	public void setGroupLocalService(GroupLocalService groupLocalService);
-
-	public void setResourceActionLocalService(
-		ResourceActionLocalService resourceActionLocalService);
-
-	public void setResourcePermissionLocalService(
-		ResourcePermissionLocalService resourcePermissionLocalService);
-
-	public void setRoleLocalService(RoleLocalService roleLocalService);
-
-	public default Filter toFilter(String filterString) {
-		return toFilter(
-			filterString, Collections.<String, List<String>>emptyMap());
-	}
-
-	public default Filter toFilter(
-		String filterString, Map<String, List<String>> multivaluedMap) {
-
-		return null;
-	}
-
-	public static class FactoryHolder {
-
-		public static volatile Factory factory;
-
-	}
-
-	@ProviderType
-	public interface Builder {
-
-		public AccountResource build();
-
-		public Builder checkPermissions(boolean checkPermissions);
-
-		public Builder httpServletRequest(
-			HttpServletRequest httpServletRequest);
-
-		public Builder httpServletResponse(
-			HttpServletResponse httpServletResponse);
-
-		public Builder preferredLocale(Locale preferredLocale);
-
-		public Builder user(com.liferay.portal.kernel.model.User user);
-
-	}
-
-	@ProviderType
-	public interface Factory {
-
-		public Builder create();
-
-	}
-
-}
