+++ modules/amf-registration/amf-registration-rest/amf-registration-rest-client/bnd.bnd
@@ -1,3 +1,3 @@
 Bundle-Name: amf-registration-rest-client
-Bundle-SymbolicName: com.liferay.amf.client
+Bundle-SymbolicName: com.liferay.amf.rest.client
 Bundle-Version: 1.0.0
