--- modules/amf-registration/amf-registration-rest/amf-registration-rest-client/src/main/java/com/liferay/amf/client/aggregation/Aggregation.java
@@ -1,25 +1,0 @@
-package com.liferay.amf.client.aggregation;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.annotation.Generated;
-
-/**
- * @author me
- * @generated
- */
-@Generated("")
-public class Aggregation {
-
-	public Map<String, String> getAggregationTerms() {
-		return _aggregationTerms;
-	}
-
-	public void setAggregationTerms(Map<String, String> aggregationTerms) {
-		_aggregationTerms = aggregationTerms;
-	}
-
-	private Map<String, String> _aggregationTerms = new HashMap<>();
-
-}
