--- modules/amf-registration/amf-registration-rest/amf-registration-rest-client/src/main/java/com/liferay/amf/client/aggregation/Facet.java
@@ -1,65 +1,0 @@
-package com.liferay.amf.client.aggregation;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.annotation.Generated;
-
-/**
- * @author me
- * @generated
- */
-@Generated("")
-public class Facet {
-
-	public Facet() {
-	}
-
-	public Facet(String facetCriteria, List<FacetValue> facetValues) {
-		_facetCriteria = facetCriteria;
-		_facetValues = facetValues;
-	}
-
-	public String getFacetCriteria() {
-		return _facetCriteria;
-	}
-
-	public List<FacetValue> getFacetValues() {
-		return _facetValues;
-	}
-
-	public void setFacetCriteria(String facetCriteria) {
-		_facetCriteria = facetCriteria;
-	}
-
-	public void setFacetValues(List<FacetValue> facetValues) {
-		_facetValues = facetValues;
-	}
-
-	public static class FacetValue {
-
-		public FacetValue() {
-		}
-
-		public FacetValue(Integer numberOfOccurrences, String term) {
-			_numberOfOccurrences = numberOfOccurrences;
-			_term = term;
-		}
-
-		public Integer getNumberOfOccurrences() {
-			return _numberOfOccurrences;
-		}
-
-		public String getTerm() {
-			return _term;
-		}
-
-		private Integer _numberOfOccurrences;
-		private String _term;
-
-	}
-
-	private String _facetCriteria;
-	private List<FacetValue> _facetValues = new ArrayList<>();
-
-}
