--- modules/amf-registration/amf-registration-rest/amf-registration-rest-client/src/main/java/com/liferay/amf/client/dto/v1_0/Account.java
@@ -1,409 +1,0 @@
-package com.liferay.amf.client.dto.v1_0;
-
-import com.liferay.amf.client.function.UnsafeSupplier;
-import com.liferay.amf.client.serdes.v1_0.AccountSerDes;
-
-import java.io.Serializable;
-
-import java.util.Objects;
-
-import javax.annotation.Generated;
-
-/**
- * @author me
- * @generated
- */
-@Generated("")
-public class Account implements Cloneable, Serializable {
-
-	public static Account toDTO(String json) {
-		return AccountSerDes.toDTO(json);
-	}
-
-	public String getAccountName() {
-		return accountName;
-	}
-
-	public void setAccountName(String accountName) {
-		this.accountName = accountName;
-	}
-
-	public void setAccountName(
-		UnsafeSupplier<String, Exception> accountNameUnsafeSupplier) {
-
-		try {
-			accountName = accountNameUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String accountName;
-
-	public String getAddress1() {
-		return address1;
-	}
-
-	public void setAddress1(String address1) {
-		this.address1 = address1;
-	}
-
-	public void setAddress1(
-		UnsafeSupplier<String, Exception> address1UnsafeSupplier) {
-
-		try {
-			address1 = address1UnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String address1;
-
-	public String getAddress2() {
-		return address2;
-	}
-
-	public void setAddress2(String address2) {
-		this.address2 = address2;
-	}
-
-	public void setAddress2(
-		UnsafeSupplier<String, Exception> address2UnsafeSupplier) {
-
-		try {
-			address2 = address2UnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String address2;
-
-	public String getBirthday() {
-		return birthday;
-	}
-
-	public void setBirthday(String birthday) {
-		this.birthday = birthday;
-	}
-
-	public void setBirthday(
-		UnsafeSupplier<String, Exception> birthdayUnsafeSupplier) {
-
-		try {
-			birthday = birthdayUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String birthday;
-
-	public String getCity() {
-		return city;
-	}
-
-	public void setCity(String city) {
-		this.city = city;
-	}
-
-	public void setCity(UnsafeSupplier<String, Exception> cityUnsafeSupplier) {
-		try {
-			city = cityUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String city;
-
-	public String getConfirmPassword() {
-		return confirmPassword;
-	}
-
-	public void setConfirmPassword(String confirmPassword) {
-		this.confirmPassword = confirmPassword;
-	}
-
-	public void setConfirmPassword(
-		UnsafeSupplier<String, Exception> confirmPasswordUnsafeSupplier) {
-
-		try {
-			confirmPassword = confirmPasswordUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String confirmPassword;
-
-	public String getEmailAddress() {
-		return emailAddress;
-	}
-
-	public void setEmailAddress(String emailAddress) {
-		this.emailAddress = emailAddress;
-	}
-
-	public void setEmailAddress(
-		UnsafeSupplier<String, Exception> emailAddressUnsafeSupplier) {
-
-		try {
-			emailAddress = emailAddressUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String emailAddress;
-
-	public String getFirstName() {
-		return firstName;
-	}
-
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
-
-	public void setFirstName(
-		UnsafeSupplier<String, Exception> firstNameUnsafeSupplier) {
-
-		try {
-			firstName = firstNameUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String firstName;
-
-	public String getGenre() {
-		return genre;
-	}
-
-	public void setGenre(String genre) {
-		this.genre = genre;
-	}
-
-	public void setGenre(
-		UnsafeSupplier<String, Exception> genreUnsafeSupplier) {
-
-		try {
-			genre = genreUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String genre;
-
-	public String getHomePhone() {
-		return homePhone;
-	}
-
-	public void setHomePhone(String homePhone) {
-		this.homePhone = homePhone;
-	}
-
-	public void setHomePhone(
-		UnsafeSupplier<String, Exception> homePhoneUnsafeSupplier) {
-
-		try {
-			homePhone = homePhoneUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String homePhone;
-
-	public String getLastName() {
-		return lastName;
-	}
-
-	public void setLastName(String lastName) {
-		this.lastName = lastName;
-	}
-
-	public void setLastName(
-		UnsafeSupplier<String, Exception> lastNameUnsafeSupplier) {
-
-		try {
-			lastName = lastNameUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String lastName;
-
-	public String getMobilePhone() {
-		return mobilePhone;
-	}
-
-	public void setMobilePhone(String mobilePhone) {
-		this.mobilePhone = mobilePhone;
-	}
-
-	public void setMobilePhone(
-		UnsafeSupplier<String, Exception> mobilePhoneUnsafeSupplier) {
-
-		try {
-			mobilePhone = mobilePhoneUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String mobilePhone;
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-	public void setPassword(
-		UnsafeSupplier<String, Exception> passwordUnsafeSupplier) {
-
-		try {
-			password = passwordUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String password;
-
-	public String getSecurityAnswer() {
-		return securityAnswer;
-	}
-
-	public void setSecurityAnswer(String securityAnswer) {
-		this.securityAnswer = securityAnswer;
-	}
-
-	public void setSecurityAnswer(
-		UnsafeSupplier<String, Exception> securityAnswerUnsafeSupplier) {
-
-		try {
-			securityAnswer = securityAnswerUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String securityAnswer;
-
-	public String getSecurityQuestion() {
-		return securityQuestion;
-	}
-
-	public void setSecurityQuestion(String securityQuestion) {
-		this.securityQuestion = securityQuestion;
-	}
-
-	public void setSecurityQuestion(
-		UnsafeSupplier<String, Exception> securityQuestionUnsafeSupplier) {
-
-		try {
-			securityQuestion = securityQuestionUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String securityQuestion;
-
-	public String getState() {
-		return state;
-	}
-
-	public void setState(String state) {
-		this.state = state;
-	}
-
-	public void setState(
-		UnsafeSupplier<String, Exception> stateUnsafeSupplier) {
-
-		try {
-			state = stateUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String state;
-
-	public String getTermsOfUse() {
-		return termsOfUse;
-	}
-
-	public void setTermsOfUse(String termsOfUse) {
-		this.termsOfUse = termsOfUse;
-	}
-
-	public void setTermsOfUse(
-		UnsafeSupplier<String, Exception> termsOfUseUnsafeSupplier) {
-
-		try {
-			termsOfUse = termsOfUseUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String termsOfUse;
-
-	@Override
-	public Account clone() throws CloneNotSupportedException {
-		return (Account)super.clone();
-	}
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof Account)) {
-			return false;
-		}
-
-		Account account = (Account)object;
-
-		return Objects.equals(toString(), account.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		return AccountSerDes.toJSON(this);
-	}
-
-}
