+++ modules/amf-registration/amf-registration-rest/amf-registration-rest-client/src/main/java/com/liferay/amf/rest/client/function/UnsafeSupplier.java
@@ -1,0 +1,15 @@
+package com.liferay.amf.rest.client.function;
+
+import javax.annotation.Generated;
+
+/**
+ * @author me
+ * @generated
+ */
+@FunctionalInterface
+@Generated("")
+public interface UnsafeSupplier<T, E extends Throwable> {
+
+	public T get() throws E;
+
+}
