+++ modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/bnd.bnd
@@ -1,6 +1,3 @@
 Bundle-Name: amf-registration-rest-impl
-Bundle-SymbolicName: com.liferay.amf.impl
+Bundle-SymbolicName: com.liferay.amf.rest.impl
 Bundle-Version: 1.0.0
-Import-Package:\
-	com.liferay.amf.dto.v1_0,\
-	com.liferay.amf.resource.v1_0
