+++ modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/build.gradle
@@ -1,11 +1,11 @@
 dependencies {
 	compile project(":modules:amf-registration:amf-registration-rest:amf-registration-rest-api")
-	compile project(":modules:amf-registration:amf-registration-api")
-	compileOnly group: "com.fasterxml.jackson.core", name: "jackson-annotations", version: "2.10.3"
+
+	compileOnly group: "com.fasterxml.jackson.core", name: "jackson-annotations", version: "2.12.4"
 	compileOnly group: "com.liferay", name: "com.liferay.petra.function"
 	compileOnly group: "com.liferay", name: "com.liferay.petra.string"
 	compileOnly group: "com.liferay", name: "com.liferay.portal.odata.api"
-	compileOnly group: "com.liferay", name: "com.liferay.portal.vulcan.api", version: "7.13.0"
+	compileOnly group: "com.liferay", name: "com.liferay.portal.vulcan.api", version: "9.3.1"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel"
 	compileOnly group: "commons-collections", name: "commons-collections"
@@ -13,12 +13,13 @@ dependencies {
 	compileOnly group: "javax.annotation", name: "javax.annotation-api", version: "1.3.2"
 	compileOnly group: "javax.validation", name: "validation-api", version: "2.0.1.Final"
 	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api"
+	compileOnly group: "javax.xml.bind", name: "jaxb-api", version: "2.3.0"
 	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
-	compileOnly group: "org.osgi", name: "org.osgi.core"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component"
+	compileOnly (group: "org.osgi", name: "org.osgi.service.component", version: "1.3.0") {force = true}
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations"
+	compileOnly group: "org.osgi", name: "org.osgi.core"
 
-	restBuilder group: "com.liferay", name: "com.liferay.portal.tools.rest.builder", version: "1.0.167"
+	restBuilder group: "com.liferay", name: "com.liferay.portal.tools.rest.builder", version: "1.0.222"
 }
 
-group = "com.liferay.amf"
+group = "com.liferay.amf.rest"
