--- modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/src/main/java/com/liferay/amf/internal/graphql/mutation/v1_0/Mutation.java
@@ -1,153 +1,0 @@
-package com.liferay.amf.internal.graphql.mutation.v1_0;
-
-import com.liferay.amf.dto.v1_0.Account;
-import com.liferay.amf.resource.v1_0.AccountResource;
-import com.liferay.petra.function.UnsafeConsumer;
-import com.liferay.petra.function.UnsafeFunction;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLField;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLName;
-
-import java.util.function.BiFunction;
-
-import javax.annotation.Generated;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
-
-import org.osgi.service.component.ComponentServiceObjects;
-
-/**
- * @author me
- * @generated
- */
-@Generated("")
-public class Mutation {
-
-	public static void setAccountResourceComponentServiceObjects(
-		ComponentServiceObjects<AccountResource>
-			accountResourceComponentServiceObjects) {
-
-		_accountResourceComponentServiceObjects =
-			accountResourceComponentServiceObjects;
-	}
-
-	@GraphQLField
-	public Account createAccount(@GraphQLName("account") Account account)
-		throws Exception {
-
-		return _applyComponentServiceObjects(
-			_accountResourceComponentServiceObjects,
-			this::_populateResourceContext,
-			accountResource -> accountResource.createAccount(account));
-	}
-
-	@GraphQLField
-	public boolean deleteAccount(@GraphQLName("accountId") String accountId)
-		throws Exception {
-
-		_applyVoidComponentServiceObjects(
-			_accountResourceComponentServiceObjects,
-			this::_populateResourceContext,
-			accountResource -> accountResource.deleteAccount(accountId));
-
-		return true;
-	}
-
-	@GraphQLField
-	public Response deleteAccountBatch(
-			@GraphQLName("callbackURL") String callbackURL,
-			@GraphQLName("object") Object object)
-		throws Exception {
-
-		return _applyComponentServiceObjects(
-			_accountResourceComponentServiceObjects,
-			this::_populateResourceContext,
-			accountResource -> accountResource.deleteAccountBatch(
-				callbackURL, object));
-	}
-
-	@GraphQLField
-	public Account updateAccount(
-			@GraphQLName("accountId") String accountId,
-			@GraphQLName("account") Account account)
-		throws Exception {
-
-		return _applyComponentServiceObjects(
-			_accountResourceComponentServiceObjects,
-			this::_populateResourceContext,
-			accountResource -> accountResource.updateAccount(
-				accountId, account));
-	}
-
-	private <T, R, E1 extends Throwable, E2 extends Throwable> R
-			_applyComponentServiceObjects(
-				ComponentServiceObjects<T> componentServiceObjects,
-				UnsafeConsumer<T, E1> unsafeConsumer,
-				UnsafeFunction<T, R, E2> unsafeFunction)
-		throws E1, E2 {
-
-		T resource = componentServiceObjects.getService();
-
-		try {
-			unsafeConsumer.accept(resource);
-
-			return unsafeFunction.apply(resource);
-		}
-		finally {
-			componentServiceObjects.ungetService(resource);
-		}
-	}
-
-	private <T, E1 extends Throwable, E2 extends Throwable> void
-			_applyVoidComponentServiceObjects(
-				ComponentServiceObjects<T> componentServiceObjects,
-				UnsafeConsumer<T, E1> unsafeConsumer,
-				UnsafeConsumer<T, E2> unsafeFunction)
-		throws E1, E2 {
-
-		T resource = componentServiceObjects.getService();
-
-		try {
-			unsafeConsumer.accept(resource);
-
-			unsafeFunction.accept(resource);
-		}
-		finally {
-			componentServiceObjects.ungetService(resource);
-		}
-	}
-
-	private void _populateResourceContext(AccountResource accountResource)
-		throws Exception {
-
-		accountResource.setContextAcceptLanguage(_acceptLanguage);
-		accountResource.setContextCompany(_company);
-		accountResource.setContextHttpServletRequest(_httpServletRequest);
-		accountResource.setContextHttpServletResponse(_httpServletResponse);
-		accountResource.setContextUriInfo(_uriInfo);
-		accountResource.setContextUser(_user);
-		accountResource.setGroupLocalService(_groupLocalService);
-		accountResource.setRoleLocalService(_roleLocalService);
-	}
-
-	private static ComponentServiceObjects<AccountResource>
-		_accountResourceComponentServiceObjects;
-
-	private AcceptLanguage _acceptLanguage;
-	private com.liferay.portal.kernel.model.Company _company;
-	private GroupLocalService _groupLocalService;
-	private HttpServletRequest _httpServletRequest;
-	private HttpServletResponse _httpServletResponse;
-	private RoleLocalService _roleLocalService;
-	private BiFunction<Object, String, Sort[]> _sortsBiFunction;
-	private UriInfo _uriInfo;
-	private com.liferay.portal.kernel.model.User _user;
-
-}
