--- modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/src/main/java/com/liferay/amf/internal/graphql/query/v1_0/Query.java
@@ -1,155 +1,0 @@
-package com.liferay.amf.internal.graphql.query.v1_0;
-
-import com.liferay.amf.dto.v1_0.Account;
-import com.liferay.amf.resource.v1_0.AccountResource;
-import com.liferay.petra.function.UnsafeConsumer;
-import com.liferay.petra.function.UnsafeFunction;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.search.filter.Filter;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLField;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLName;
-import com.liferay.portal.vulcan.pagination.Page;
-
-import java.util.Map;
-import java.util.function.BiFunction;
-
-import javax.annotation.Generated;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import javax.ws.rs.core.UriInfo;
-
-import org.osgi.service.component.ComponentServiceObjects;
-
-/**
- * @author me
- * @generated
- */
-@Generated("")
-public class Query {
-
-	public static void setAccountResourceComponentServiceObjects(
-		ComponentServiceObjects<AccountResource>
-			accountResourceComponentServiceObjects) {
-
-		_accountResourceComponentServiceObjects =
-			accountResourceComponentServiceObjects;
-	}
-
-	/**
-	 * Invoke this method with the command line:
-	 *
-	 * curl -H 'Content-Type: text/plain; charset=utf-8' -X 'POST' 'http://localhost:8080/o/graphql' -d $'{"query": "query {allAccount(pageNumber: ___, pageSize: ___){items {__}, page, pageSize, totalCount}}"}' -u 'test@liferay.com:test'
-	 */
-	@GraphQLField(description = "Return all accounts")
-	public AccountPage allAccount(
-			@GraphQLName("pageNumber") Integer pageNumber,
-			@GraphQLName("pageSize") Integer pageSize)
-		throws Exception {
-
-		return _applyComponentServiceObjects(
-			_accountResourceComponentServiceObjects,
-			this::_populateResourceContext,
-			accountResource -> new AccountPage(
-				accountResource.getAllAccount(pageNumber, pageSize)));
-	}
-
-	/**
-	 * Invoke this method with the command line:
-	 *
-	 * curl -H 'Content-Type: text/plain; charset=utf-8' -X 'POST' 'http://localhost:8080/o/graphql' -d $'{"query": "query {account(accountId: ___){firstName, lastName, emailAddress, accountName, genre, birthday, password, confirmPassword, homePhone, mobilePhone, address1, address2, city, state, securityQuestion, securityAnswer, termsOfUse}}"}' -u 'test@liferay.com:test'
-	 */
-	@GraphQLField
-	public Account account(@GraphQLName("accountId") String accountId)
-		throws Exception {
-
-		return _applyComponentServiceObjects(
-			_accountResourceComponentServiceObjects,
-			this::_populateResourceContext,
-			accountResource -> accountResource.getAccount(accountId));
-	}
-
-	@GraphQLName("AccountPage")
-	public class AccountPage {
-
-		public AccountPage(Page accountPage) {
-			actions = accountPage.getActions();
-
-			items = accountPage.getItems();
-			lastPage = accountPage.getLastPage();
-			page = accountPage.getPage();
-			pageSize = accountPage.getPageSize();
-			totalCount = accountPage.getTotalCount();
-		}
-
-		@GraphQLField
-		protected Map<String, Map> actions;
-
-		@GraphQLField
-		protected java.util.Collection<Account> items;
-
-		@GraphQLField
-		protected long lastPage;
-
-		@GraphQLField
-		protected long page;
-
-		@GraphQLField
-		protected long pageSize;
-
-		@GraphQLField
-		protected long totalCount;
-
-	}
-
-	private <T, R, E1 extends Throwable, E2 extends Throwable> R
-			_applyComponentServiceObjects(
-				ComponentServiceObjects<T> componentServiceObjects,
-				UnsafeConsumer<T, E1> unsafeConsumer,
-				UnsafeFunction<T, R, E2> unsafeFunction)
-		throws E1, E2 {
-
-		T resource = componentServiceObjects.getService();
-
-		try {
-			unsafeConsumer.accept(resource);
-
-			return unsafeFunction.apply(resource);
-		}
-		finally {
-			componentServiceObjects.ungetService(resource);
-		}
-	}
-
-	private void _populateResourceContext(AccountResource accountResource)
-		throws Exception {
-
-		accountResource.setContextAcceptLanguage(_acceptLanguage);
-		accountResource.setContextCompany(_company);
-		accountResource.setContextHttpServletRequest(_httpServletRequest);
-		accountResource.setContextHttpServletResponse(_httpServletResponse);
-		accountResource.setContextUriInfo(_uriInfo);
-		accountResource.setContextUser(_user);
-		accountResource.setGroupLocalService(_groupLocalService);
-		accountResource.setRoleLocalService(_roleLocalService);
-	}
-
-	private static ComponentServiceObjects<AccountResource>
-		_accountResourceComponentServiceObjects;
-
-	private AcceptLanguage _acceptLanguage;
-	private com.liferay.portal.kernel.model.Company _company;
-	private BiFunction<Object, String, Filter> _filterBiFunction;
-	private GroupLocalService _groupLocalService;
-	private HttpServletRequest _httpServletRequest;
-	private HttpServletResponse _httpServletResponse;
-	private RoleLocalService _roleLocalService;
-	private BiFunction<Object, String, Sort[]> _sortsBiFunction;
-	private UriInfo _uriInfo;
-	private com.liferay.portal.kernel.model.User _user;
-
-}
