--- modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/src/main/java/com/liferay/amf/internal/graphql/servlet/v1_0/ServletDataImpl.java
@@ -1,53 +1,0 @@
-package com.liferay.amf.internal.graphql.servlet.v1_0;
-
-import com.liferay.amf.internal.graphql.mutation.v1_0.Mutation;
-import com.liferay.amf.internal.graphql.query.v1_0.Query;
-import com.liferay.amf.resource.v1_0.AccountResource;
-import com.liferay.portal.vulcan.graphql.servlet.ServletData;
-
-import javax.annotation.Generated;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.ComponentServiceObjects;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceScope;
-
-/**
- * @author me
- * @generated
- */
-@Component(immediate = true, service = ServletData.class)
-@Generated("")
-public class ServletDataImpl implements ServletData {
-
-	@Activate
-	public void activate(BundleContext bundleContext) {
-		Mutation.setAccountResourceComponentServiceObjects(
-			_accountResourceComponentServiceObjects);
-
-		Query.setAccountResourceComponentServiceObjects(
-			_accountResourceComponentServiceObjects);
-	}
-
-	@Override
-	public Mutation getMutation() {
-		return new Mutation();
-	}
-
-	@Override
-	public String getPath() {
-		return "/amf-registration-rest-graphql/v1_0";
-	}
-
-	@Override
-	public Query getQuery() {
-		return new Query();
-	}
-
-	@Reference(scope = ReferenceScope.PROTOTYPE_REQUIRED)
-	private ComponentServiceObjects<AccountResource>
-		_accountResourceComponentServiceObjects;
-
-}
