--- modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/src/main/java/com/liferay/amf/internal/jaxrs/application/AmfRegistrationRestApplication.java
@@ -1,24 +1,0 @@
-package com.liferay.amf.internal.jaxrs.application;
-
-import javax.annotation.Generated;
-
-import javax.ws.rs.core.Application;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author me
- * @generated
- */
-@Component(
-	property = {
-		"liferay.jackson=false",
-		"osgi.jaxrs.application.base=/amf-registration-rest",
-		"osgi.jaxrs.extension.select=(osgi.jaxrs.name=Liferay.Vulcan)",
-		"osgi.jaxrs.name=AmfRegistrationRest"
-	},
-	service = Application.class
-)
-@Generated("")
-public class AmfRegistrationRestApplication extends Application {
-}
