+++ modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/src/main/java/com/liferay/amf/rest/internal/resource/v1_0/AccountResourceImpl.java
@@ -1,0 +1,16 @@
+package com.liferay.amf.rest.internal.resource.v1_0;
+
+import com.liferay.amf.rest.resource.v1_0.AccountResource;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ServiceScope;
+
+/**
+ * @author me
+ */
+@Component(
+	properties = "OSGI-INF/liferay/rest/v1_0/account.properties",
+	scope = ServiceScope.PROTOTYPE, service = AccountResource.class
+)
+public class AccountResourceImpl extends BaseAccountResourceImpl {
+}
