+++ modules/amf-registration/amf-registration-rest/amf-registration-rest-impl/src/main/java/com/liferay/amf/rest/internal/resource/v1_0/OpenAPIResourceImpl.java
@@ -1,0 +1,75 @@
+package com.liferay.amf.rest.internal.resource.v1_0;
+
+import com.liferay.portal.vulcan.resource.OpenAPIResource;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.annotations.info.License;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import javax.annotation.Generated;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.UriInfo;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author me
+ * @generated
+ */
+@Component(
+	properties = "OSGI-INF/liferay/rest/v1_0/openapi.properties",
+	service = OpenAPIResourceImpl.class
+)
+@Generated("")
+@OpenAPIDefinition(
+	info = @Info(description = "The Sample API made with rest builder. A Java client JAR is available for use with the group ID 'com.liferay', artifact ID 'com.liferay.sample.client', and version '1.0.0'.. A Java client JAR is available for use with the group ID 'com.liferay', artifact ID 'com.liferay.amf.client', and version '1.0.0'.. A Java client JAR is available for use with the group ID 'com.liferay', artifact ID 'com.liferay.amf.rest.client', and version '1.0.0'.", license = @License(name = "Apache 2.0", url = "http://www.apache.org/licenses/LICENSE-2.0.html"), title = "Acme Movie Fanatics (AMF)  API", version = "v1.0")
+)
+@Path("/v1.0")
+public class OpenAPIResourceImpl {
+
+	@GET
+	@Path("/openapi.{type:json|yaml}")
+	@Produces({MediaType.APPLICATION_JSON, "application/yaml"})
+	public Response getOpenAPI(@PathParam("type") String type)
+		throws Exception {
+
+		try {
+			Class<? extends OpenAPIResource> clazz =
+				_openAPIResource.getClass();
+
+			clazz.getMethod(
+				"getOpenAPI", Set.class, String.class, UriInfo.class);
+		}
+		catch (NoSuchMethodException noSuchMethodException) {
+			return _openAPIResource.getOpenAPI(_resourceClasses, type);
+		}
+
+		return _openAPIResource.getOpenAPI(_resourceClasses, type, _uriInfo);
+	}
+
+	@Reference
+	private OpenAPIResource _openAPIResource;
+
+	@Context
+	private UriInfo _uriInfo;
+
+	private final Set<Class<?>> _resourceClasses = new HashSet<Class<?>>() {
+		{
+			add(AccountResourceImpl.class);
+
+			add(OpenAPIResourceImpl.class);
+		}
+	};
+
+}
