--- modules/amf-registration/amf-registration-rest/amf-registration-rest-test/src/testIntegration/java/com/liferay/amf/resource/v1_0/test/AccountResourceTest.java
@@ -1,14 +1,0 @@
-package com.liferay.amf.resource.v1_0.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-
-import org.junit.Ignore;
-import org.junit.runner.RunWith;
-
-/**
- * @author me
- */
-@Ignore
-@RunWith(Arquillian.class)
-public class AccountResourceTest extends BaseAccountResourceTestCase {
-}
